Finish new task in first level

+++ 1.JavaSyntax/src/com/javarush/task/task10/task1006/Solution.java
@@ -6,12 +6,12 @@ package com.javarush.task.task10.task1006;
 
 public class Solution {
     public static void main(String[] args) {
-        short b = (short) 45;
+        short b = 45;
         char c = (short) 'c';
         short s = (short) 1005.22;
-        int i = (short) 150000;
-        float f = (short) 4.10f;
-        double d = (short) 1.256d;
+        int i = 150000;
+        float f = 4.10f;
+        double d = 1.256d;
         double result = (f * b) + (i / c) - (d * s) + 562.78d;
         System.out.println("result: " + result);
     }
